+++ tools/diff-cli/api/diff-cli.api
@@ -1,0 +1,4 @@
+public final class io/matthewnelson/diff/cli/MainKt {
+	public static final fun main ([Ljava/lang/String;)V
+}
+
