+++ tools/diff-cli/core/.gitignore
