+++ tools/diff-cli/core/build.gradle.kts
