+++ tools/diff-cli/core/src/commonMain/kotlin/io/matthewnelson/diff/core/internal/-BufferedSink.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+@file:Suppress("KotlinRedundantDiagnosticSuppress")
+
+package io.matthewnelson.diff.core.internal
+
+import okio.BufferedSink
+
+@Suppress("NOTHING_TO_INLINE")
+internal inline fun BufferedSink.writeNewLine() {
+    writeUtf8("\n")
+}
+
+@Suppress("NOTHING_TO_INLINE")
+internal inline fun BufferedSink.writeLineBreak() {
+    writeNewLine()
+    writeUtf8(LINE_BREAK)
+    writeNewLine()
+}
