+++ tools/diff-cli/core/src/commonMain/kotlin/io/matthewnelson/diff/core/internal/InternalDiffApi.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package io.matthewnelson.diff.core.internal
+
+/**
+ * Denotes an api as internal, and subject to change at any time.
+ *
+ * Any usage of a declaration annotated with [InternalDiffApi]
+ * must be accepted by annotating that usage with the [OptIn]
+ * annotation, e.g @OptIn(InternalDiffApi::class), or by using
+ * the following compiler argument:
+ *
+ *   -Xopt-in=io.matthewnelson.diff.core.internal.InternalDiffApi
+ * */
+@RequiresOptIn
+@MustBeDocumented
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.ANNOTATION_CLASS,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.LOCAL_VARIABLE,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.CONSTRUCTOR,
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.TYPEALIAS
+)
+@Retention(AnnotationRetention.BINARY)
+public annotation class InternalDiffApi
