+++ tools/diff-cli/core/src/jvmMain/kotlin/io/matthewnelson/diff/core/internal/-JvmPlatform.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+@file:Suppress("KotlinRedundantDiagnosticSuppress")
+
+package io.matthewnelson.diff.core.internal
+
+import okio.FileSystem
+
+@InternalDiffApi
+@Suppress("NOTHING_TO_INLINE")
+public actual inline fun FileSystem.Companion.system(): FileSystem = SYSTEM
